fix two separate bugs around RGB aux channels
First off, green was not defined properly and was instead defined the
same as a main channel (leading to the behaviour described in the issue
at https://github.com/ToyKeeper/anduril/pull/33#issuecomment-1891341281

Second, fix the actual bug causing aux to not work for blinky modes.
Specifically, set_level needs to be patched so that instead of (or as
well as, doesn't really matter) calling set_level_zero(), it will also
check for if the active channel is an RGB channel and if so them clear
RGB state.