Fix issue #107 blocking channel-mode events past 3H on d4k-3ch.
If a light has a custom 3H function specified for a channel, it would
always call the custom function, not just during 3H events. It would
only return if the event was handled by the custom function, but would
fall out of the if/else otherwise, preventing later events from being
handled.

This prevented access to the 9H channel config in HSV mode, for
instance.